fix: fix broken website CSS by refactoring styled-components SSR logic (#257)
A number of mistakes in the code resulted in the page not being properly
server-side rendered in the Next.JS static export. This means all DOM
hydration and stylesheet injection was being performed client-side,
which defeats the purpose of a static export / SSR with NextJS. This
also seemed to be non-deterministic in terms of the order stylesheets
were injected into the `<head>` client-side, which resulted in buggy /
unstyled rendering for some users.

The following changes seems to have fixed the issue: 
- Refactor global styles to use `createGlobalStyle` functionality from
`styled-components`
- Ensure `layout.tsx` is server side rendered by moving all context
providers to a separate client component.
- Remove the `Suspense` boundary that was required for using the NextJS
`useSearchParams` hook. Using native browser APIs
(`window.location.search`) works fine here, so no need for a suspense
boundary with this custom hook.